backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
openai==1.12.0
google-generativeai==0.3.1
pymongo==4.6.1
redis==5.0.1
email-validator==2.1.0
requests==2.31.0
werkzeug==3.0.1
